lease: moved addrconf updater to addrconf.[ch]
