+++ src/ifconfig.c
@@ -78,6 +78,7 @@
 #include "sysfs.h"
 #include "kernel.h"
 #include "appconfig.h"
+#include "addrconf.h"
 #include "process.h"
 #include "debug.h"
 #include "modprobe.h"
@@ -534,8 +535,8 @@ ni_system_interface_link_monitor(ni_netdev_t *dev)
 /*
  * system interface lease updater actions
  */
-static int
-__ni_addrconf_action_mtu_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_mtu_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 
@@ -548,8 +549,8 @@ __ni_addrconf_action_mtu_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 
 }
 
-static int
-__ni_addrconf_action_addrs_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_addrs_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 	int res = -1;
@@ -566,7 +567,7 @@ __ni_addrconf_action_addrs_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 }
 
 static int
-__ni_addrconf_action_addrs_verify_check(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+ni_addrconf_action_addrs_verify_check(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	unsigned int duplicates = 0;
 	unsigned int tentative = 0;
@@ -658,7 +659,7 @@ __ni_addrconf_action_addrs_verify_check(ni_netdev_t *dev, ni_addrconf_lease_t *l
 }
 
 static int
-__ni_addrconf_action_addrs_verify_loop(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+ni_addrconf_action_addrs_verify_loop(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 	unsigned int loops = 50;
@@ -673,7 +674,7 @@ __ni_addrconf_action_addrs_verify_loop(ni_netdev_t *dev, ni_addrconf_lease_t *le
 		if (!nc || (res = __ni_system_refresh_interface_addrs(nc, dev)) < 0)
 			return res;
 
-		if ((res = __ni_addrconf_action_addrs_verify_check(dev, lease)) <= 0)
+		if ((res = ni_addrconf_action_addrs_verify_check(dev, lease)) <= 0)
 			return res;
 
 		/* In case the client is configured to ignore link-up
@@ -693,7 +694,7 @@ __ni_addrconf_action_addrs_verify_loop(ni_netdev_t *dev, ni_addrconf_lease_t *le
 }
 
 static int
-__ni_addrconf_action_addrs_verify_timed(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+ni_addrconf_action_addrs_verify_timed(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 	int res = -1;
@@ -702,7 +703,7 @@ __ni_addrconf_action_addrs_verify_timed(ni_netdev_t *dev, ni_addrconf_lease_t *l
 		return res;
 
 	/* updater is executed on address updates removing tentative flags */
-	if ((res = __ni_addrconf_action_addrs_verify_check(dev, lease)) == 0)
+	if ((res = ni_addrconf_action_addrs_verify_check(dev, lease)) == 0)
 		return res;
 
 	/* This should not happen for auto6 addresses as the
@@ -724,19 +725,19 @@ __ni_addrconf_action_addrs_verify_timed(ni_netdev_t *dev, ni_addrconf_lease_t *l
 	return res;
 }
 
-static int
-__ni_addrconf_action_addrs_verify(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_addrs_verify(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_addrconf_updater_t *updater = lease->updater;
 
 	if (updater && timerisset(&updater->started))
-		return __ni_addrconf_action_addrs_verify_timed(dev, lease);
+		return ni_addrconf_action_addrs_verify_timed(dev, lease);
 	else
-		return __ni_addrconf_action_addrs_verify_loop(dev, lease);
+		return ni_addrconf_action_addrs_verify_loop(dev, lease);
 }
 
 static int
-__ni_addrconf_action_verify_address_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+ni_addrconf_action_verify_address_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	unsigned int duplicates = 0;
 	unsigned int verified = 0;
@@ -768,8 +769,8 @@ __ni_addrconf_action_verify_address_apply(ni_netdev_t *dev, ni_addrconf_lease_t 
 	return 0;
 }
 
-static int
-__ni_addrconf_action_routes_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_routes_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 	int res;
@@ -789,8 +790,8 @@ __ni_addrconf_action_routes_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 	return 0;
 }
 
-static int
-__ni_addrconf_action_system_update(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_system_update(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	lease->update &= ni_config_addrconf_update_mask(lease->type, lease->family);
 
@@ -801,30 +802,30 @@ __ni_addrconf_action_system_update(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 	return 0;
 }
 
-static int
-__ni_addrconf_action_verify_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_verify_apply(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
-	return __ni_addrconf_action_verify_address_apply(dev, lease);
+	return ni_addrconf_action_verify_address_apply(dev, lease);
 }
 
-int
-__ni_addrconf_action_write_lease(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_write_lease(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	lease->state = NI_ADDRCONF_STATE_GRANTED;
 	ni_addrconf_lease_file_write(dev->name, lease);
 	return 0;
 }
 
-int
-__ni_addrconf_action_remove_lease(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_remove_lease(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	lease->state = NI_ADDRCONF_STATE_RELEASED;
 	ni_addrconf_lease_file_write(dev->name, lease);
 	return 0;
 }
 
-static int
-__ni_addrconf_action_addrs_remove(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_addrs_remove(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 	int res;
@@ -841,8 +842,8 @@ __ni_addrconf_action_addrs_remove(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 	return 0;
 }
 
-static int
-__ni_addrconf_action_routes_remove(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_routes_remove(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 	int res;
@@ -862,8 +863,8 @@ __ni_addrconf_action_routes_remove(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 	return 0;
 }
 
-static int
-__ni_addrconf_action_mtu_restore(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
+extern int
+ni_addrconf_action_mtu_restore(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 {
 	ni_netconfig_t *nc = ni_global_state_handle(0);
 
@@ -874,378 +875,6 @@ __ni_addrconf_action_mtu_restore(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
 	 */
 	return 0;
 
-}
-
-struct ni_addrconf_action {
-	int		(*func)(ni_netdev_t *dev, ni_addrconf_lease_t *lease);
-	const char *	info;
-};
-
-static const ni_addrconf_action_t	updater_applying_common[] = {
-	{ __ni_addrconf_action_mtu_apply,	"adjusting mtu"		},
-	{ __ni_addrconf_action_addrs_apply,	"applying addresses"	},
-	{ __ni_addrconf_action_addrs_verify,	"verifying adressses"	},
-	{ __ni_addrconf_action_routes_apply,	"applying routes"	},
-	{ __ni_addrconf_action_system_update,	"applying system config"},
-	{ __ni_addrconf_action_verify_apply,	"verifying apply state" },
-	{ NULL,	NULL }
-};
-
-static const ni_addrconf_action_t	updater_removing_common[] = {
-	{ __ni_addrconf_action_addrs_remove,	"removing addresses"	},
-	{ __ni_addrconf_action_routes_remove,	"removing routes"	},
-	{ __ni_addrconf_action_system_update,	"removing system config"},
-	{ __ni_addrconf_action_mtu_restore,	"reverting mtu change"	},
-	{ NULL,	NULL }
-};
-
-static const ni_addrconf_action_t	updater_applying_auto6[] = {
-	{ __ni_addrconf_action_addrs_verify,	"verifying adressses"	},
-	{ __ni_addrconf_action_write_lease,	"writing lease file"   },
-	{ __ni_addrconf_action_system_update,	"applying system config"},
-	{ NULL, NULL }
-};
-
-static const ni_addrconf_action_t	updater_removing_auto6[] = {
-	{ __ni_addrconf_action_system_update,	"applying system config"},
-	{ __ni_addrconf_action_remove_lease,	"removing lease file"   },
-	{ NULL, NULL }
-};
-
-static ni_addrconf_updater_t *
-ni_addrconf_updater_new(const ni_addrconf_action_t *action, const ni_netdev_t *dev, ni_event_t event)
-{
-	ni_addrconf_updater_t *updater;
-
-	updater = xcalloc(1, sizeof(*updater));
-	if (updater) {
-		updater->event  = event;
-		updater->action = action;
-		ni_timer_get_time(&updater->started);
-		if (dev)
-			ni_netdev_ref_set(&updater->device, dev->name, dev->link.ifindex);
-	}
-	return updater;
-}
-
-ni_addrconf_updater_t *
-ni_addrconf_updater_new_applying(ni_addrconf_lease_t *lease, const ni_netdev_t *dev, ni_event_t event)
-{
-	if (!lease)
-		return NULL;
-
-	ni_addrconf_updater_free(&lease->updater);
-	if (lease->family == AF_INET6 && lease->type == NI_ADDRCONF_AUTOCONF) {
-		lease->updater = ni_addrconf_updater_new(updater_applying_auto6, dev, event);
-	} else {
-		lease->updater = ni_addrconf_updater_new(updater_applying_common, dev, event);
-	}
-	return lease->updater;
-}
-
-ni_addrconf_updater_t *
-ni_addrconf_updater_new_removing(ni_addrconf_lease_t *lease, const ni_netdev_t *dev, ni_event_t event)
-{
-	if (!lease)
-		return NULL;
-
-	ni_addrconf_updater_free(&lease->updater);
-	if (lease->family == AF_INET6 && lease->type == NI_ADDRCONF_AUTOCONF) {
-		lease->updater = ni_addrconf_updater_new(updater_removing_auto6, dev, event);
-	} else {
-		lease->updater = ni_addrconf_updater_new(updater_removing_common, dev, event);
-	}
-	return lease->updater;
-}
-
-static int
-ni_addrconf_updater_action_call(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
-{
-	ni_addrconf_updater_t *updater;
-	struct timeval now, delta;
-	int res = 0;
-
-	if (!dev || !lease)
-		return 0;
-
-	while ((updater = lease->updater) != NULL) {
-		if (!updater->action || !updater->action->func) {
-			ni_addrconf_updater_free(&lease->updater);
-			break;
-		}
-
-		if (!timerisset(&updater->astart))
-			ni_timer_get_time(&updater->astart);
-
-		res = updater->action->func(dev, lease);
-
-		ni_timer_get_time(&now);
-		if (timercmp(&now, &updater->astart, >))
-			timersub(&now, &updater->astart, &delta);
-		else
-			timerclear(&delta);
-
-		if (updater->action->info) {
-			ni_debug_verbose(NI_LOG_DEBUG, NI_TRACE_IFCONFIG,
-					"%s: %s for %s:%s lease in state %s: %s [%s %ldm%ld.%03lds]",
-					dev->name, updater->action->info,
-					ni_addrfamily_type_to_name(lease->family),
-					ni_addrconf_type_to_name(lease->type),
-					ni_addrconf_state_to_name(lease->state),
-					(res < 0  ? "failure"  :
-					 res > 0  ? "deferred" : "success"),
-					(res == 1 ? "since"    : "after"),
-					delta.tv_sec / 60, delta.tv_sec % 60,
-					delta.tv_usec / 1000);
-		}
-		if (res != 1)
-			timerclear(&updater->astart);
-		if (res != 0)
-			break;
-		updater->action++;
-	}
-	return res;
-}
-
-int
-ni_addrconf_updater_execute(ni_netdev_t *dev, ni_addrconf_lease_t *lease)
-{
-	ni_addrconf_updater_t *updater;
-	struct timeval pre, now, delta;
-	ni_event_t event;
-	int ret = 0;
-
-	if (!dev || !lease || !(updater = lease->updater))
-		return -1;
-
-	if (updater->timer)
-		ni_timer_cancel(updater->timer);
-	updater->timer = NULL;
-
-	event = updater->event;
-	pre = updater->started;
-	ret = ni_addrconf_updater_action_call(dev, lease);
-
-	if (ret > 0) {
-		struct timeval now;
-
-		ni_timer_get_time(&now);
-		if (updater->deadline &&
-		    !ni_lifetime_left(updater->deadline, &updater->started, &now)) {
-			lease->state = NI_ADDRCONF_STATE_FAILED;
-			ni_debug_ifconfig("%s: %s:%s lease update deadline reached (state %s)",
-					dev->name,
-					ni_addrfamily_type_to_name(lease->family),
-					ni_addrconf_type_to_name(lease->type),
-					ni_addrconf_state_to_name(lease->state));
-			ret = -1;
-		} else
-		if (!ni_addrconf_updater_background(updater, updater->timeout)) {
-			ni_debug_ifconfig("%s: unable to background %s:%s lease update (state %s)",
-					dev->name,
-					ni_addrfamily_type_to_name(lease->family),
-					ni_addrconf_type_to_name(lease->type),
-					ni_addrconf_state_to_name(lease->state));
-			ret = -1;
-		} else {
-			return ret;
-		}
-	} else {
-		ni_timer_get_time(&now);
-		if (timercmp(&now, &pre, >)) {
-			timersub(&now, &pre, &delta);
-		} else {
-			timerclear(&delta);
-		}
-		ni_debug_ifconfig("%s: updater for lease %s:%s in state %s finished: %s [%ldm%ld.%03lds]",
-				dev->name,
-				ni_addrfamily_type_to_name(lease->family),
-				ni_addrconf_type_to_name(lease->type),
-				ni_addrconf_state_to_name(lease->state),
-				ret < 0 ? "failure" : "success",
-				delta.tv_sec / 60, delta.tv_sec % 60,
-				delta.tv_usec / 1000);
-	}
-
-	if (ret == 0) {
-		if (lease->old) {
-			ni_addrconf_lease_free(lease->old);
-			lease->old = NULL;
-		}
-		ni_addrconf_updater_free(&lease->updater);
-
-		switch (lease->state) {
-		case NI_ADDRCONF_STATE_GRANTED:
-		case NI_ADDRCONF_STATE_APPLYING:
-			lease->state = NI_ADDRCONF_STATE_GRANTED;
-			if (event != NI_EVENT_ADDRESS_ACQUIRED)
-				ret = -1;
-			else
-			if (ni_addrconf_flag_bit_is_set(lease->flags, NI_ADDRCONF_FLAGS_PRIMARY))
-				ni_objectmodel_addrconf_fallback_action(dev, event, lease->family, NULL);
-			break;
-
-		case NI_ADDRCONF_STATE_RELEASED:
-		case NI_ADDRCONF_STATE_RELEASING:
-			lease->state = NI_ADDRCONF_STATE_RELEASED;
-			if (event != NI_EVENT_ADDRESS_RELEASED)
-				ret = -1;
-			break;
-
-		case NI_ADDRCONF_STATE_FAILED:
-			if (event != NI_EVENT_ADDRESS_RELEASED &&
-			    event != NI_EVENT_ADDRESS_LOST)
-				ret = -1;
-			break;
-
-		case NI_ADDRCONF_STATE_REQUESTING:
-			/* reenter requesting on partial success */
-			if (event == NI_EVENT_ADDRESS_ACQUIRED)
-				event = NI_EVENT_ADDRESS_DEFERRED;
-			else
-				ret = -1;
-			break;
-
-		default:
-			lease->state = NI_ADDRCONF_STATE_FAILED;
-			event = __NI_EVENT_MAX;
-			break;
-		}
-
-		if (ret == 0 && event < __NI_EVENT_MAX) {
-			ni_debug_ifconfig("%s: %s:%s lease updated (state %s), sending %s event",
-					dev->name,
-					ni_addrfamily_type_to_name(lease->family),
-					ni_addrconf_type_to_name(lease->type),
-					ni_addrconf_state_to_name(lease->state),
-					ni_objectmodel_event_to_signal(event));
-
-			ni_objectmodel_addrconf_send_event(dev, event, &lease->uuid);
-			if (event == NI_EVENT_ADDRESS_RELEASED) {
-				if (ni_addrconf_flag_bit_is_set(lease->flags, NI_ADDRCONF_FLAGS_FALLBACK))
-					ni_objectmodel_addrconf_fallback_action(dev, event, lease->family, lease);
-				else
-					ni_netdev_unset_lease(dev, lease->family, lease->type);
-			}
-		}
-		if (ret == 0)
-			return ret;
-	}
-
-	if (ret < 0) {
-		if (lease->old) {
-			ni_addrconf_lease_free(lease->old);
-			lease->old = NULL;
-		}
-		ni_addrconf_updater_free(&lease->updater);
-
-		/* aborted by dad, dhcp6 supplicant will decline */
-		if (lease->family == AF_INET6 && lease->state == NI_ADDRCONF_STATE_REQUESTING)
-			return 0;
-
-		lease->state = NI_ADDRCONF_STATE_FAILED;
-		ni_objectmodel_addrconf_send_event(dev, NI_EVENT_ADDRESS_LOST, &lease->uuid);
-	}
-	return ret;
-}
-
-static void
-ni_addrconf_updater_timer_call(void *user_data, const ni_timer_t *timer)
-{
-	ni_addrconf_updater_t *updater = user_data;
-	ni_addrconf_lease_t *lease;
-	ni_netconfig_t *nc;
-	ni_netdev_t *dev;
-
-	if (!updater || updater->timer != timer)
-		return;
-
-	updater->timer = NULL;
-
-	if (!(nc = ni_global_state_handle(0)))
-		return;
-
-	if (!(dev = ni_netdev_by_index(nc, updater->device.index)))
-		return;
-
-	for (lease = dev->leases; lease; lease = lease->next) {
-		if (lease->updater == updater) {
-			ni_addrconf_updater_execute(dev, lease);
-			return;
-		}
-	}
-}
-
-ni_bool_t
-ni_addrconf_updater_background(ni_addrconf_updater_t *updater, unsigned int delay)
-{
-	unsigned long timeout = 0;
-
-	if (!updater)
-		return FALSE;
-
-	if (!updater->timeout)
-		updater->timeout = 1000;
-
-	updater->jitter.max = 100;
-	if (delay > 1000)
-		updater->jitter.min = 0 - updater->jitter.max;
-	else
-		updater->jitter.min = 0;
-	timeout = ni_timeout_randomize(delay, &updater->jitter);
-
-	if (updater->timer != NULL)
-		updater->timer = ni_timer_rearm(updater->timer, timeout);
-
-	if (updater->timer == NULL)
-		updater->timer = ni_timer_register(timeout,
-				ni_addrconf_updater_timer_call, updater);
-
-	return updater->timer != NULL;
-}
-
-void
-ni_addrconf_updater_set_data(ni_addrconf_updater_t *updater, void *user_data,
-				ni_addrconf_updater_cleanup_t *cleanup)
-{
-	if (updater) {
-		if (updater->user_data && updater->cleanup) {
-			void *data = updater->user_data;
-			updater->user_data = NULL;
-			updater->cleanup(data);
-		}
-		updater->user_data = user_data;
-		updater->cleanup = cleanup;
-	}
-}
-
-void *
-ni_addrconf_updater_get_data(ni_addrconf_updater_t *updater,
-				ni_addrconf_updater_cleanup_t *cleanup)
-{
-	if (updater && updater->cleanup == cleanup)
-		return updater->user_data;
-	return NULL;
-}
-
-static inline void
-ni_addrconf_updater_destroy(ni_addrconf_updater_t *updater)
-{
-	if (updater->timer)
-		ni_timer_cancel(updater->timer);
-	updater->timer = NULL;
-	ni_addrconf_updater_set_data(updater, NULL, NULL);
-	ni_netdev_ref_destroy(&updater->device);
-}
-
-void
-ni_addrconf_updater_free(ni_addrconf_updater_t **updater)
-{
-	if (updater && *updater) {
-		ni_addrconf_updater_destroy(*updater);
-		free(*updater);
-		*updater = NULL;
-	}
 }
 
 /*
